--- src/controllers/classicformController/index.ts
@@ -1,2 +1,0 @@
-export * from './classicformController';
-export * from './types';
