--- src/controllers/paperformController/helpers.ts
@@ -1,26 +1,0 @@
-import DOMPurify from 'isomorphic-dompurify';
-import qs from 'qs';
-import { head, pipe } from 'ramda';
-import { PaperFormParams } from './types';
-
-export const stringify = (params: Record<string, string>) => qs.stringify(params, { indices: false });
-const escape = (val?: string): string => (typeof val === 'string' ? DOMPurify.sanitize(val) : '');
-const listSanitizer = (v: string): string[] =>
-  v.length > 0 ? (Array.from(v.matchAll(/[^\r\n]+/g), head) as string[]) : [];
-
-export const checks = {
-  listCheck: pipe(escape, listSanitizer),
-  escape: (val: string) => escape(val),
-};
-
-export const stringifiers = {
-  journalForm({ bibstem, year, volume, page }: PaperFormParams) {
-    const q = [];
-    bibstem.length > 0 && q.push(`bibstem:${bibstem}`);
-    year.length > 0 && q.push(`year:${year}`);
-    volume.length > 0 && q.push(`volume:${volume}`);
-    page.length > 0 && q.push(`page:${page}`);
-
-    return stringify({ q: q.join(' ') });
-  },
-};
