--- src/controllers/paperformController/index.ts
@@ -1,2 +1,0 @@
-export * from './paperformController';
-export * from './types';
